% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dewp.R
\name{dewp3}
\alias{dewp3}
\title{Probability mass function of the three-parameter EWP}
\usage{
dewp3(x, lambda, beta1, beta2, sum_limit = max(x) * 3)
}
\arguments{
\item{x}{vector of (positive integer) quantiles.}

\item{lambda}{centrality parameter}

\item{beta1}{lower-tail dispersion parameter}

\item{beta2}{upper tail dispersion parameter}

\item{sum_limit}{summation limit for the normalizing factor}
}
\value{
a vector of probabilities
}
\description{
Probability mass function of the three-parameter EWP
}
