% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dewp3_cpp}
\alias{dewp3_cpp}
\title{Probability mass function of the three-parameter EWP}
\usage{
dewp3_cpp(x, lambda, beta1, beta2, sum_limit)
}
\arguments{
\item{x}{vector of (positive integer) quantiles.}

\item{lambda}{centrality parameter}

\item{beta1}{lower-tail dispersion parameter}

\item{beta2}{upper tail dispersion parameter}

\item{sum_limit}{summation limit for the normalizing factor}
}
\value{
a probability mass
}
\description{
Probability mass function of the three-parameter EWP
}
