% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewp_reg.R
\name{fitted.ewp}
\alias{fitted.ewp}
\title{Extract fitted values}
\usage{
\method{fitted}{ewp}(object, ...)
}
\arguments{
\item{object}{an object of class ewp}

\item{...}{ignored}
}
\value{
a vector of fitted values on the response scale
}
\description{
Extract fitted values
}
