% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{linnet}
\alias{linnet}
\title{Linnet clutch sizes}
\format{
A data frame with 5414 rows and 3 variables:
\describe{
  \item{eggs}{clutch size}
  \item{cov1}{a synthetic random noise covariate}
  \item{cov2}{a synthetic covariate that is positively correlated with the outcome}
  }
}
\source{
Ridout & Besbeas 2004, P. Boersch-Supan
}
\usage{
linnet
}
\description{
A dataset containing the clutch sizes for linnet, recreated from Ridout & Besbeas 2004
}
\keyword{datasets}
