% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewp_reg.R
\name{logLik.ewp}
\alias{logLik.ewp}
\title{Extract log likelihood}
\usage{
\method{logLik}{ewp}(object, ...)
}
\arguments{
\item{object}{an object of class ewp}

\item{...}{ignored}
}
\value{
a numeric
}
\description{
Extract log likelihood
}
