% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewp_reg.R
\name{predict.ewp}
\alias{predict.ewp}
\title{Predict from fitted model}
\usage{
\method{predict}{ewp}(object, newdata, type = c("response"), na.action = na.pass, ...)
}
\arguments{
\item{object}{ewp model object}

\item{newdata}{optional data.frame}

\item{type}{character; default="response", no other type implemented}

\item{na.action}{defaults to na.pass()}

\item{...}{ignored}
}
\value{
a vector of predictions
}
\description{
Predict from fitted model
}
