% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewp_reg.R
\name{print.ewp}
\alias{print.ewp}
\title{Print ewp model object}
\usage{
\method{print}{ewp}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{ewp model object}

\item{digits}{digits to print}

\item{...}{ignored}
}
\description{
Print ewp model object
}
