% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewp_reg.R
\name{print.summary.ewp}
\alias{print.summary.ewp}
\title{Print ewp model summary}
\usage{
\method{print}{summary.ewp}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{ewp model summary}

\item{digits}{number of digits to print}

\item{...}{additional arguments to printCoefmat()}
}
\value{
printout of the summary object
}
\description{
Print ewp model summary
}
