% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dewp.R
\name{rewp3}
\alias{rewp3}
\title{Random samples from the three-parameter EWP}
\usage{
rewp3(n, lambda, beta1, beta2, sum_limit = 30)
}
\arguments{
\item{n}{number of observations}

\item{lambda}{centrality parameter}

\item{beta1}{lower-tail dispersion parameter}

\item{beta2}{upper tail dispersion parameter}

\item{sum_limit}{summation limit for the normalizing factor}
}
\value{
random deviates from the EWP_3 distribution
}
\description{
Random samples from the three-parameter EWP
}
