% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewp_reg.R
\name{simulate.ewp}
\alias{simulate.ewp}
\title{simulate from fitted model}
\usage{
\method{simulate}{ewp}(object, nsim = 1, ...)
}
\arguments{
\item{object}{ewp model object}

\item{nsim}{number of response vectors to simulate. Defaults to 1.}

\item{...}{ignored}
}
\value{
a data frame with `nsim` columns.
}
\description{
simulate from fitted model
}
