% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewp_reg.R
\name{vcov.ewp}
\alias{vcov.ewp}
\title{Extract estimated variance-covariance matrix}
\usage{
\method{vcov}{ewp}(object, ...)
}
\arguments{
\item{object}{an object of class ewp}

\item{...}{ignored}
}
\value{
a matrix
}
\description{
Extract estimated variance-covariance matrix
}
